requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
affine==2.3.1
    # via
    #   odc-geo
    #   rasterio
appdirs==1.4.4
    # via geocube
attrs==22.1.0
    # via
    #   fiona
    #   rasterio
azure-core==1.26.0
    # via
    #   azure-storage-blob
    #   msrest
azure-storage-blob==12.14.1
    # via -r requirements.in
bokeh==2.4.3
    # via dask
cachetools==5.2.0
    # via odc-geo
certifi==2022.9.24
    # via
    #   fiona
    #   msrest
    #   pyproj
    #   rasterio
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   click-plugins
    #   cligj
    #   distributed
    #   fiona
    #   geocube
    #   planetary-computer
    #   rasterio
click-plugins==1.1.1
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
cloudpickle==2.2.0
    # via
    #   dask
    #   distributed
cryptography==38.0.1
    # via azure-storage-blob
dask[array,complete]==2022.10.0
    # via
    #   -r requirements.in
    #   distributed
    #   stackstac
distributed==2022.10.0
    # via dask
fiona==1.8.22
    # via geopandas
fsspec==2022.10.0
    # via dask
geocube==0.3.3
    # via -r requirements.in
geopandas==0.12.0
    # via
    #   -r requirements.in
    #   geocube
heapdict==1.0.1
    # via zict
idna==3.4
    # via requests
isodate==0.6.1
    # via msrest
jinja2==3.1.2
    # via
    #   bokeh
    #   dask
    #   distributed
locket==1.0.0
    # via
    #   distributed
    #   partd
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.4
    # via distributed
msrest==0.7.1
    # via azure-storage-blob
munch==2.5.0
    # via fiona
numpy==1.23.4
    # via
    #   bokeh
    #   dask
    #   geocube
    #   odc-geo
    #   pandas
    #   pygeos
    #   rasterio
    #   scipy
    #   snuggs
    #   xarray
oauthlib==3.2.2
    # via requests-oauthlib
odc-geo==0.3.2
    # via geocube
packaging==21.3
    # via
    #   bokeh
    #   dask
    #   distributed
    #   geopandas
    #   rioxarray
    #   xarray
pandas==1.5.1
    # via
    #   dask
    #   geopandas
    #   xarray
partd==1.3.0
    # via dask
pillow==9.2.0
    # via bokeh
planetary-computer==0.4.7
    # via -r requirements.in
psutil==5.9.3
    # via distributed
pycparser==2.21
    # via cffi
pydantic[dotenv]==1.10.2
    # via planetary-computer
pygeos==0.13
    # via -r requirements.in
pyparsing==3.0.9
    # via
    #   packaging
    #   snuggs
pyproj==3.4.0
    # via
    #   geocube
    #   geopandas
    #   odc-geo
    #   rioxarray
    #   stackstac
pystac==1.6.1
    # via
    #   planetary-computer
    #   pystac-client
pystac-client==0.5.1
    # via
    #   -r requirements.in
    #   planetary-computer
python-dateutil==2.8.2
    # via
    #   pandas
    #   pystac
    #   pystac-client
python-dotenv==0.21.0
    # via pydantic
pytz==2022.5
    # via
    #   pandas
    #   planetary-computer
pyyaml==6.0
    # via
    #   bokeh
    #   dask
    #   distributed
rasterio==1.3.3
    # via
    #   geocube
    #   rioxarray
    #   stackstac
requests==2.28.1
    # via
    #   azure-core
    #   msrest
    #   planetary-computer
    #   pystac-client
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via msrest
rioxarray==0.12.2
    # via
    #   -r requirements.in
    #   geocube
scipy==1.9.3
    # via geocube
shapely==1.8.5.post1
    # via
    #   geopandas
    #   odc-geo
six==1.16.0
    # via
    #   azure-core
    #   fiona
    #   isodate
    #   munch
    #   python-dateutil
snuggs==1.4.7
    # via rasterio
sortedcontainers==2.4.0
    # via distributed
stackstac==0.4.3
    # via -r requirements.in
tblib==1.7.0
    # via distributed
toolz==0.12.0
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via
    #   bokeh
    #   distributed
typing-extensions==4.4.0
    # via
    #   azure-core
    #   bokeh
    #   pydantic
urllib3==1.26.12
    # via
    #   distributed
    #   requests
xarray==2022.10.0
    # via
    #   -r requirements.in
    #   geocube
    #   rioxarray
    #   stackstac
zict==2.2.0
    # via distributed

# The following packages are considered to be unsafe in a requirements file:
# setuptools
